Move repeated fields to field package
