+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -21,11 +21,11 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.field.Email;
+import seedu.address.model.field.Name;
+import seedu.address.model.field.Phone;
+import seedu.address.model.field.Price;
 import seedu.address.model.property.Address;
-import seedu.address.model.property.Email;
-import seedu.address.model.property.Name;
-import seedu.address.model.property.Phone;
-import seedu.address.model.property.Price;
 import seedu.address.model.property.Property;
 import seedu.address.model.property.Seller;
 import seedu.address.model.tag.Tag;
