+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.model.Model;
-import seedu.address.model.property.NameContainsKeywordsPredicate;
+import seedu.address.model.field.NameContainsKeywordsPredicate;
 
 /**
  * Finds and lists all properties in address book whose name contains any of the argument keywords.
