+++ src/main/java/seedu/address/model/property/Property.java
@@ -7,6 +7,10 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+import seedu.address.model.field.Email;
+import seedu.address.model.field.Name;
+import seedu.address.model.field.Phone;
+import seedu.address.model.field.Price;
 import seedu.address.model.tag.Tag;
 
 /**
