+++ src/test/java/seedu/address/model/property/PhoneTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.field.Phone;
+
 public class PhoneTest {
 
     @Test
