+++ src/test/java/seedu/address/model/property/PriceTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.field.Price;
+
 public class PriceTest {
     @Test
     public void constructor_null_throwsNullPointerException() {
