+++ src/test/java/seedu/address/storage/JsonAdaptedPropertyTest.java
@@ -12,11 +12,11 @@ import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.model.field.Email;
+import seedu.address.model.field.Name;
+import seedu.address.model.field.Phone;
+import seedu.address.model.field.Price;
 import seedu.address.model.property.Address;
-import seedu.address.model.property.Email;
-import seedu.address.model.property.Name;
-import seedu.address.model.property.Phone;
-import seedu.address.model.property.Price;
 import seedu.address.model.property.Seller;
 
 public class JsonAdaptedPropertyTest {
