+++ src/test/java/seedu/address/testutil/EditPropertyDescriptorBuilder.java
@@ -5,11 +5,11 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import seedu.address.logic.commands.EditCommand;
+import seedu.address.model.field.Email;
+import seedu.address.model.field.Name;
+import seedu.address.model.field.Phone;
+import seedu.address.model.field.Price;
 import seedu.address.model.property.Address;
-import seedu.address.model.property.Email;
-import seedu.address.model.property.Name;
-import seedu.address.model.property.Phone;
-import seedu.address.model.property.Price;
 import seedu.address.model.property.Property;
 import seedu.address.model.property.Seller;
 import seedu.address.model.tag.Tag;
