+++ src/test/resources/view/UiPartTest/validFileWithFxRoot.fxml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import seedu.address.ui.TestFxmlObject?>
 <fx:root type="seedu.address.ui.TestFxmlObject" xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml">
+         xmlns:fx="http://javafx.com/fxml">
     <text>Hello World!</text>
 </fx:root>
